sniffer: remove old application
