--- ng_sniffer/Makefile
@@ -1,20 +1,0 @@
-# Set the name of your application:
-APPLICATION = ng_sniffer
-
-# If no BOARD is found in the environment, use this default:
-BOARD ?= native
-
-# This has to be the absolute path to the RIOT base directory:
-RIOTBASE ?= $(CURDIR)/../../RIOT
-
-# Define modules that are used
-USEMODULE += ng_netif_default
-USEMODULE += auto_init_ng_netif
-USEMODULE += shell
-USEMODULE += shell_commands
-USEMODULE += ps
-
-# Change this to 0 show compiler invocation lines by default:
-QUIET ?= 1
-
-include $(RIOTBASE)/Makefile.include
